updated game code & drew a boss
I have updated the game code to use variables for colors. I also drew a boss. the boss does not do anything yet. I did build a game finished screen that currently shows up after the 15th level is complete. the Idea will be to have the boss be the 15th level and if you beat him the game will be over.